Remove handleErrors and add logErrors
