+++ packages/express/src/index.ts
@@ -1,2 +1,2 @@
 export { getCallback } from './get-callback';
-export { handleErrors } from './handle-errors';
+export { logErrors } from './log-errors';
