+++ packages/express/src/log-errors.ts
@@ -1,0 +1,8 @@
+export function logErrors(enabled = true, logFn = console.error) {
+  return (err, req, res, next) => {
+    if (enabled) {
+      logFn(err);
+    }
+    next(err);
+  };
+}
